TimingTuple       WARNING TimingTuple:: Tuple 'TimingTuple' 'unsignedlong' has different sizes on 32/64 bit systems. Casting 'EventInSequence' to 'unsigned long long'
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@164044 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: a2a54b1ad0c3756a54c24474a6e1ff0de42cae67

+++ Rec/Rec/RecAlgs/src/TimingTuple.h
@@ -39,7 +39,7 @@ private:
 
   ISequencerTimerTool* m_timerTool; ///< timer tool
   int m_timer ; ///< timer index
-  unsigned long int m_evtCounter;
+  unsigned long long m_evtCounter;
   std::string m_recSummaryLoc;
 
 private:
